+++ PythonProject/DeepFace Tests/deepface_test_results/matches.csv
@@ -1,0 +1,3 @@
+id,img_name,model_name,detector_backend,aligned,l2_normalized,search_method,target_x,target_y,target_w,target_h,threshold,distance_metric,distance,confidence
+4,/Users/dilloncoker/Desktop/spring_26/machine_learning/ml_project_files/imdb_crop/61/nm0861361_rm575703808_1969-11-20_2004.jpg,VGG-Face,opencv,True,False,exact,253,68,127,127,0.68,cosine,0.49948368170839574,69.36
+11,/Users/dilloncoker/Desktop/spring_26/machine_learning/ml_project_files/imdb_crop/61/nm0861361_rm3628181504_1969-11-20_2011.jpg,VGG-Face,opencv,True,False,exact,253,68,127,127,0.68,cosine,0.6792527863247033,51.0
